refactor: move matrix extensions to another class and add type extensions for generic type name resolving
